football-prediction-app/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
beautifulsoup4==4.12.2
lxml==4.9.3
joblib==1.3.2
xgboost==2.0.3
lightgbm==4.2.0
statsmodels==0.14.1
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
APScheduler==3.10.4
click==8.1.7
cryptography==41.0.7
pytz==2023.3.post1
psutil==5.9.6